algebra: double to float

+++ mat2.h
@@ -5,18 +5,18 @@ typedef union u_mat2 mat2;
 
 union u_mat2 {
     struct {
-        double m00, m10,
-               m01, m11;
+        float m00, m10,
+              m01, m11;
     };
 
-    double ptr[2][2];
+    float ptr[2][2];
 };
 
 extern mat2 mat2Zero(void);
 
 extern mat2 mat2Identity(void);
 
-extern mat2 mat2Scalar(const double);
+extern mat2 mat2Scalar(const float);
 
 extern mat2 mat2Add(const mat2, const mat2);
 
@@ -24,9 +24,9 @@ extern mat2 mat2Sub(const mat2, const mat2);
 
 extern mat2 mat2Mul(const mat2, const mat2);
 
-extern mat2 mat2Mulf(const mat2, const double);
+extern mat2 mat2Mulf(const mat2, const float);
 
-extern double mat2Det(const mat2);
+extern float mat2Det(const mat2);
 
 # if defined (ALGEBRA_IMPLEMENTATION)
 
@@ -40,7 +40,7 @@ extern mat2 mat2Identity(void) {
                       0.0, 1.0 }} );
 }
 
-extern mat2 mat2Scalar(const double s) {
+extern mat2 mat2Scalar(const float s) {
     return ((mat2) {{ 1.0 * s, 0.0,
                       0.0,     1.0 * s }} );
 }
