+++ mat4.h
@@ -9,18 +9,18 @@ typedef union u_mat4 mat4;
 
 union u_mat4 {
     struct {
-        double m00, m10, m20, m30,
-               m01, m11, m21, m31,
-               m02, m12, m22, m32,
-               m03, m13, m23, m33;
+        float m00, m10, m20, m30,
+              m01, m11, m21, m31,
+              m02, m12, m22, m32,
+              m03, m13, m23, m33;
     };
 
-    double ptr[4][4];
+    float ptr[4][4];
 };
 
 extern mat4 mat4Zero(void);
 
-extern mat4 mat4Init(const double);
+extern mat4 mat4Init(const float);
 
 extern mat4 mat4Add(const mat4, const mat4);
 
@@ -28,15 +28,15 @@ extern mat4 mat4Sub(const mat4, const mat4);
 
 extern mat4 mat4Mul(const mat4, const mat4);
 
-extern mat4 mat4Mulf(const mat4, const double);
+extern mat4 mat4Mulf(const mat4, const float);
 
-extern double mat4Det(const mat4);
+extern float mat4Det(const mat4);
 
-extern mat4 mat4Frust(const double, const double, const double, const double, const double, const double);
+extern mat4 mat4Frust(const float, const float, const float, const float, const float, const float);
 
-extern mat4 mat4Ortho(const double, const double, const double, const double, const double, const double);
+extern mat4 mat4Ortho(const float, const float, const float, const float, const float, const float);
 
-extern mat4 mat4Persp(const double, const double, const double, const double);
+extern mat4 mat4Persp(const float, const float, const float, const float);
 
 # if defined (ALGEBRA_IMPLEMENTATION)
 
@@ -47,7 +47,7 @@ extern mat4 mat4Zero(void) {
                       0.0, 0.0, 0.0, 0.0 }} );
 }
 
-extern mat4 mat4Init(const double s) {
+extern mat4 mat4Init(const float s) {
     return ((mat4) {{ 1.0 * s, 0.0,     0.0,     0.0,
                       0.0,     1.0 * s, 0.0,     0.0,
                       0.0,     0.0,     1.0 * s, 0.0,
@@ -90,15 +90,15 @@ extern mat4 mat4Mul(const mat4 a, const mat4 b) {
                       a.m03 * b.m30 + a.m13 * b.m31 + a.m23 * b.m32 + a.m33 * b.m33 }} );
 }
 
-extern mat4 mat4Mulf(const mat4 a, const double f) {
+extern mat4 mat4Mulf(const mat4 a, const float f) {
     return ((mat4) {{ a.m00 * f, a.m01 * f, a.m02 * f, a.m03 * f,
                       a.m10 * f, a.m11 * f, a.m12 * f, a.m13 * f,
                       a.m20 * f, a.m21 * f, a.m22 * f, a.m23 * f,
                       a.m30 * f, a.m31 * f, a.m32 * f, a.m33 * f }} );
 }
 
-extern double mat4Det(const mat4 a) {
-    double result = 0.0;
+extern float mat4Det(const mat4 a) {
+    float result = 0.0;
 
     result += a.m00 * mat3Det((mat3) {{ a.m11, a.m12, a.m13,
                                         a.m21, a.m22, a.m23,
@@ -119,7 +119,7 @@ extern double mat4Det(const mat4 a) {
     return (result);
 }
 
-extern mat4 mat4Frust(const double left, const double right, const double top, const double down, const double near, const double far) {
+extern mat4 mat4Frust(const float left, const float right, const float top, const float down, const float near, const float far) {
     mat4 mat = mat4Zero();
     mat.m00  = (near * 2.0) / (right - left);
     mat.m11  = (near * 2.0) / (top   - down);
@@ -131,7 +131,7 @@ extern mat4 mat4Frust(const double left, const double right, const double top, c
     return (mat);
 }
 
-extern mat4 mat4Ortho(const double left, const double right, const double top, const double down, const double near, const double far) {
+extern mat4 mat4Ortho(const float left, const float right, const float top, const float down, const float near, const float far) {
     mat4 mat = mat4Zero();
     mat.m00 =  2.0 / (right - left);
     mat.m11 = -2.0 / (top   - down);
@@ -143,9 +143,9 @@ extern mat4 mat4Ortho(const double left, const double right, const double top, c
     return (mat);
 }
 
-extern mat4 mat4Persp(const double fieldOfView, const double aspect, const double near, const double far) {
-    double top   = near * tan(fieldOfView * 0.5);
-    double right = top * aspect;
+extern mat4 mat4Persp(const float fieldOfView, const float aspect, const float near, const float far) {
+    float top   = near * tan(fieldOfView * 0.5);
+    float right = top * aspect;
     return (mat4Frust(-right, right, top, -top, near, far));
 }
